feat: extract RTMP server into standalone process and Docker container
Move LiveStreamingServerNet + FFmpeg relay/recording machinery from the
main Thiccdal bot into a self-contained Thiccdal.RtmpServer ASP.NET Core
service. The bot becomes a thin HTTP+SignalR client; all streaming logic
lives in the new service.

Key changes:
- Add Thiccdal.RtmpServer project with REST API (/api/config, /api/start,
  /api/stop, /api/status), SignalR events hub (/hubs/events), API-key
  middleware, and a Dockerfile exposing ports 8080 and 1935
- Add docker-compose.rtmp.yml for standalone deployment with volume mounts
  for recordings (./recordings:/app/recordings) and BRB slate (./brb:/app/brb)
- Replace Thiccdal.Streaming server-side services with RtmpServerClient
  (HTTP+SignalR), RemoteStreamingService, and RemoteRtmpFanoutService
- Add Infrastructure types: IRtmpServerClient, RtmpServerOptions,
  RtmpServerConfigurationPush, RtmpRelayDestinationPush,
  RtmpServerStatusResponse, RtmpServerEvent/EventType,
  IRtmpServerConfigurationHolder
- Add EF Core migration AddRtmpServerConnection (RtmpServerUrl, RtmpServerApiKey
  columns on RestreamConfiguration)
- Update RestreamRuntimeService to push config before Start, gate CanStart
  on RTMP server connectivity, and configure client on every BuildState call
- Add OBS RTMP settings helper panel in RestreamPanel.razor (derived from
  the ingest URL, no manual lookup needed)
- Register rtmp-server in Aspire AppHost; wire RtmpServer__BaseUrl env var
- Move StreamingServiceTests + RtmpFanoutServiceTests to Thiccdal.Data.Tests
  with updated constructors and fakes; extract IRtmpEventPublisher interface
